// services/121-service/src/payments/fsp-integration/commercial-bank-ethiopia/dto/commercial-bank-ethiopia-response.dto.ts
export interface CommercialBankEthiopiaResponse {
  resultCode: string;
  resultDescription: string;
  transactionId: string;
  successIndicator: string;
  FundsTtransferType: string;
  transactionType: string;
  debitAcctNo: string;
  currencyMktDr: string;
  debitCurrency: string;
  debitAmount: string;
  debitValueDate: string;
  debitTheirRef: {
    reloadable: string[];
    refundable: string[];
  };
  creditAcctNo: string;
  currencyMktCr: string;
  creditCurrency: string;
  creditValueDate: string;
  treasuryRate: string;
  processingDate: string;
  gOrderingCust: { orderingCust: string };
  chargeComDisplay: string;
  commissionCode: string;
  chargeCode: string;
  baseCurrency: string;
  profitCenterCust: string;
  returnToDept: string;
  fedFunds: string;
  positionType: string;
  amountDebited: string;
  amountCredited: string;
  customerRate: string;
  gDeliveryOutRef: { deliveryOutRef: string[] };
  creditCompCode: string;
  debitCompCode: string;
  locAmtDebited: string;
  locAmtCredited: string;
  custGroupLevel: string;
  debitCustomer: string;
  creditCustomer: string;
  drAdviceReqDyn: string;
  crAdviceReqDyn: string;
  chargedCustomer: string;
  totRecComm: string;
  totRecCommLcl: string;
  totRecChg: string;
  totRecChgLcl: string;
  rateFixing: string;
  totRecChgCrCcy: string;
  totSndChgCrCcy: string;
  authDate: string;
  roundType: string;
  gStmtNos: { stmtNos: string[] };
  currNo: string;
  gInputter: { inputter: string };
  gDateTime: { dateTime: string };
  authoriser: string;
  coCode: string;
  deptCode: string;
  remName: string;
  benName: string;
}